+++ src/m365/spfx/commands/spfx-doctor.ts
@@ -437,6 +437,21 @@ class SpfxDoctorCommand extends AnonymousCommand {
         range: '^4',
         fix: 'npm i -g yo@4'
       }
+    },
+    '1.16.0-beta.1': {
+      gulpCli: {
+        range: '^1 || ^2',
+        fix: 'npm i -g gulp-cli@2'
+      },
+      node: {
+        range: '^16.13',
+        fix: 'Install Node.js v16.13 or higher'
+      },
+      sp: SharePointVersion.SPO,
+      yo: {
+        range: '^4',
+        fix: 'npm i -g yo@4'
+      }
     }
   };